+++ packages/ibm-products/src/components/ButtonSetWithOverflow/ButtonSetWithOverflow.stories.js
@@ -22,6 +22,7 @@ import styles from './_storybook-styles.scss';
 export default {
   title: getStoryTitle(ButtonSetWithOverflow.displayName),
   component: ButtonSetWithOverflow,
+  tags: ['autodocs'],
   argTypes: {
     containerWidth: {
       control: { type: 'range', min: 20, max: 800, step: 10 },