+++ packages/ibm-products/src/components/MultiAddSelect/MultiAddSelect.stories.js
@@ -12,22 +12,23 @@ import {
   prepareStory,
 } from '../../global/js/utils/story-helper';
 import { MultiAddSelect } from '.';
-import mdx from './MultiAddSelect.mdx';
 import { Button } from '@carbon/react';
 // import { action } from '@storybook/addon-actions';
 import image from '../UserProfileImage/headshot.jpg'; // cspell:disable-line
 import { Group, Document } from '@carbon/react/icons';
 
 import { pkg } from '../../settings';
+import DocsPage from './MultiAddSelect.docs-page';
 const blockClass = `${pkg.prefix}--add-select__meta-panel`;
 
 export default {
   title: getStoryTitle(MultiAddSelect.displayName),
   component: MultiAddSelect,
+  tags: ['autodocs'],
   parameters: {
     // styles,
     docs: {
-      page: mdx,
+      page: DocsPage,
     },
   },
 };
@@ -101,8 +102,8 @@ const defaultProps = {
   filterByLabel: 'Filter',
 };
 
-const Template = (args) => {
-  const [open, setOpen] = useState(true);
+const Template = (args, context) => {
+  const [open, setOpen] = useState(context?.viewMode !== 'docs');
   return (
     <>
       <MultiAddSelect {...args} open={open} onClose={() => setOpen(false)} />