+++ packages/ibm-products/src/components/StatusIcon/StatusIcon.stories.js
@@ -8,16 +8,18 @@
 import React from 'react';
 
 import { StatusIcon } from '.';
-import mdx from './StatusIcon.mdx';
+
 // import styles from './_storybook-styles.scss'; // import storybook which includes component and additional storybook styles
 import {
   getStoryTitle,
   prepareStory,
 } from '../../global/js/utils/story-helper';
+import DocsPage from './StatusIcon.docs-page';
 
 export default {
   title: getStoryTitle(StatusIcon.displayName),
   component: StatusIcon,
+  tags: ['autodocs'],
   argTypes: {
     kind: {
       control: {
@@ -53,7 +55,7 @@ export default {
   parameters: {
     // styles,
     docs: {
-      page: mdx,
+      page: DocsPage,
     },
   },
 };