+++ config/stylelint-config-ibm-products/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.38](https://github.com/carbon-design-system/ibm-products/compare/stylelint-config-ibm-products@0.0.37...stylelint-config-ibm-products@0.0.38) (2023-06-13)
+
+**Note:** Version bump only for package stylelint-config-ibm-products
+
+
+
+
+
 ## [0.0.37](https://github.com/carbon-design-system/ibm-products/compare/stylelint-config-ibm-products@0.0.36...stylelint-config-ibm-products@0.0.37) (2023-06-05)